catch last lints for cyclo and magic numbers

+++ pkg/v1/auth/csp/token.go
@@ -133,9 +133,9 @@ var DefaultTimeout = 30
 // IsExpired checks for the token expiry and returns true if the token has expired else will return false
 func IsExpired(tokenExpiry time.Time) bool {
 	// refresh at half token life
-	two := int64(2)
+	two := 2
 	now := time.Now().Unix()
-	halfDur := -time.Duration((tokenExpiry.Unix()-now)/two) * time.Second
+	halfDur := -time.Duration((tokenExpiry.Unix()-now)/int64(two)) * time.Second
 	return tokenExpiry.Add(halfDur).Unix() < now
 }
